kexec: move KexecLoad to linux
KexecLoad loads bzimages only and is Linux-specific. It should be in the
Linux-specific package.
